requirements.txt
aiosmtpd==1.4.6
alembic==1.16.2
atpublic==6.0.1
attrs==25.3.0
babel==2.17.0
blinker==1.9.0
certifi==2025.6.15
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
elastic-transport==8.17.1
elasticsearch==8.12.0
email_validator==2.2.0
Flask==3.1.1
flask-babel==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-Moment==1.0.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
WTForms==3.2.1
